requirements.txt
asgiref==3.7.2
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
redis==5.0.0
sqlparse==0.4.4
